+++ float/src/repr.rs
@@ -629,7 +629,6 @@ impl<const B: Word> Repr<B> {
     ///
     /// This method is intended for static creation macros.
     #[doc(hidden)]
-    #[rustversion::since(1.64)]
     #[inline]
     pub const unsafe fn from_static_words(
         sign: Sign,