+++ integer/src/bits.rs
@@ -60,12 +60,8 @@ impl UBig {
     /// assert_eq!(UBig::ZERO.trailing_zeros(), None);
     /// ```
     ///
-    /// # Availability
-    ///
-    /// Const since Rust 1.64
-    #[rustversion::attr(since(1.64), const)]
     #[inline]
-    pub fn trailing_zeros(&self) -> Option<usize> {
+    pub const fn trailing_zeros(&self) -> Option<usize> {
         self.repr().trailing_zeros()
     }
 
@@ -85,12 +81,8 @@ impl UBig {
     /// assert_eq!(UBig::ZERO.trailing_ones(), Some(0));
     /// ```
     ///
-    /// # Availability
-    ///
-    /// Const since Rust 1.64
-    #[rustversion::attr(since(1.64), const)]
     #[inline]
-    pub fn trailing_ones(&self) -> Option<usize> {
+    pub const fn trailing_ones(&self) -> Option<usize> {
         Some(self.repr().trailing_ones())
     }
 
@@ -228,12 +220,8 @@ impl IBig {
     /// assert_eq!(IBig::ZERO.trailing_zeros(), None);
     /// ```
     ///
-    /// # Availability
-    ///
-    /// Const since Rust 1.64
-    #[rustversion::attr(since(1.64), const)]
     #[inline]
-    pub fn trailing_zeros(&self) -> Option<usize> {
+    pub const fn trailing_zeros(&self) -> Option<usize> {
         self.as_sign_repr().1.trailing_zeros()
     }
 
@@ -254,11 +242,7 @@ impl IBig {
     /// assert_eq!(IBig::NEG_ONE.trailing_ones(), None);
     /// ```
     ///
-    /// # Availability
-    ///
-    /// Const since Rust 1.64
-    #[rustversion::attr(since(1.64), const)]
-    pub fn trailing_ones(&self) -> Option<usize> {
+    pub const fn trailing_ones(&self) -> Option<usize> {
         let (sign, repr) = self.as_sign_repr();
         match sign {
             Positive => Some(repr.trailing_ones()),