+++ src/lib.rs
@@ -46,7 +46,6 @@ macro_rules! ubig {
 }
 
 #[macro_export]
-#[rustversion::since(1.64)]
 #[doc = include_str!("macro-docs/static_ubig.md")]
 macro_rules! static_ubig {
     ($($t:tt)+) => {
@@ -63,7 +62,6 @@ macro_rules! ibig {
 }
 
 #[macro_export]
-#[rustversion::since(1.64)]
 #[doc = include_str!("macro-docs/static_ibig.md")]
 macro_rules! static_ibig {
     ($($t:tt)+) => {
@@ -80,7 +78,6 @@ macro_rules! fbig {
 }
 
 #[macro_export]
-#[rustversion::since(1.64)]
 #[doc = include_str!("macro-docs/static_fbig.md")]
 macro_rules! static_fbig {
     ($($t:tt)+) => {
@@ -97,7 +94,6 @@ macro_rules! dbig {
 }
 
 #[macro_export]
-#[rustversion::since(1.64)]
 #[doc = include_str!("macro-docs/static_dbig.md")]
 macro_rules! static_dbig {
     ($($t:tt)+) => {
@@ -114,7 +110,6 @@ macro_rules! rbig {
 }
 
 #[macro_export]
-#[rustversion::since(1.64)]
 #[doc = include_str!("macro-docs/static_rbig.md")]
 macro_rules! static_rbig {
     ($($t:tt)+) => {
@@ -131,7 +126,6 @@ macro_rules! cbig {
 }
 
 #[macro_export]
-#[rustversion::since(1.64)]
 #[doc = include_str!("macro-docs/static_cbig.md")]
 macro_rules! static_cbig {
     ($($t:tt)+) => {